Refactor browser snapshot preview loading
