+++ src/zivo/services/browser_snapshot.py
@@ -2,18 +2,38 @@
 
 from __future__ import annotations
 
-import shutil
-import subprocess
 import threading
 from collections import OrderedDict
 from dataclasses import dataclass, field, replace
 from pathlib import Path
 from time import sleep
-from typing import Literal, Mapping, Protocol
+from typing import Mapping, Protocol
 
 from zivo.adapters import DirectoryReader, LocalFilesystemAdapter
 from zivo.archive_utils import is_supported_archive_path
 from zivo.services import ArchiveListService, LiveArchiveListService
+from zivo.services.previews import (
+    DEFAULT_IMAGE_PREVIEW_COLUMNS,
+    PREVIEW_PERMISSION_DENIED_MESSAGE,
+    TEXT_PREVIEW_MAX_BYTES,
+    ChafaImagePreviewLoader,
+    ContextPreviewState,
+    DocumentPreviewLoader,
+    FilePreviewState,
+    GrepContextCacheKey,
+    ImagePreviewLoader,
+    PandocDocumentPreviewLoader,
+    _build_grep_context_cache_key,
+    _build_text_preview_cache_key,
+    _load_grep_context_preview,
+    _load_text_preview,
+)
+from zivo.services.previews import (
+    PREVIEW_UNSUPPORTED_MESSAGE as _preview_unsupported_message,
+)
+from zivo.services.previews import (
+    preview_max_bytes_from_kib as _preview_max_bytes_from_kib,
+)
 from zivo.state.models import (
     AppState,
     BrowserSnapshot,
@@ -36,180 +56,8 @@ from zivo.windows_paths import (
 DEFAULT_DIRECTORY_CACHE_CAPACITY = 64
 DEFAULT_TEXT_PREVIEW_CACHE_CAPACITY = 128
 DEFAULT_GREP_CONTEXT_CACHE_CAPACITY = 128
-TEXT_PREVIEW_MAX_BYTES = 64 * 1024
-DEFAULT_IMAGE_PREVIEW_COLUMNS = 80
-IMAGE_PREVIEW_EXTENSIONS = frozenset(
-    {".avif", ".bmp", ".gif", ".jpeg", ".jpg", ".png", ".svg", ".tif", ".tiff", ".webp"}
-)
-PDF_PREVIEW_EXTENSIONS = frozenset({".pdf"})
-OFFICE_PREVIEW_EXTENSIONS = frozenset({".docx", ".xlsx", ".pptx"})
-TEXT_PREVIEW_EXTENSIONS = frozenset(
-    {
-        ".adoc",
-        ".ada",
-        ".adb",
-        ".ads",
-        ".asm",
-        ".avsc",
-        ".bat",
-        ".bib",
-        ".c",
-        ".capnp",
-        ".cbl",
-        ".cc",
-        ".cfg",
-        ".cljs",
-        ".clj",
-        ".cmd",
-        ".cob",
-        ".conf",
-        ".config",
-        ".containerfile",
-        ".compose",
-        ".cpp",
-        ".cr",
-        ".cql",
-        ".css",
-        ".css.map",
-        ".csv",
-        ".cypher",
-        ".d",
-        ".dart",
-        ".diff",
-        ".dockerfile",
-        ".edn",
-        ".elm",
-        ".erl",
-        ".ex",
-        ".exs",
-        ".f",
-        ".f90",
-        ".fish",
-        ".geojson",
-        ".go",
-        ".gql",
-        ".gradle",
-        ".groovy",
-        ".h",
-        ".har",
-        ".hcl",
-        ".hrl",
-        ".hpp",
-        ".html",
-        ".htmx",
-        ".ics",
-        ".ini",
-        ".java",
-        ".js.map",
-        ".js",
-        ".jl",
-        ".json",
-        ".jsonl",
-        ".jsx",
-        ".jsx.map",
-        ".hs",
-        ".ksh",
-        ".kt",
-        ".kts",
-        ".kube",
-        ".latex",
-        ".less",
-        ".log",
-        ".lua",
-        ".m",
-        ".md",
-        ".mjs.map",
-        ".make",
-        ".mk",
-        ".ml",
-        ".mli",
-        ".mm",
-        ".mysql",
-        ".ndjson",
-        ".nim",
-        ".nomad",
-        ".opts",
-        ".org",
-        ".pas",
-        ".pcss",
-        ".postcss",
-        ".pp",
-        ".prop",
-        ".properties",
-        ".proto",
-        ".ps1",
-        ".psql",
-        ".psv",
-        ".py",
-        ".patch",
-        ".rej",
-        ".rb",
-        ".ron",
-        ".rst",
-        ".rs",
-        ".s",
-        ".sass",
-        ".scala",
-        ".scss",
-        ".sh",
-        ".srt",
-        ".sv",
-        ".svh",
-        ".swift",
-        ".svelte",
-        ".sql",
-        ".tcl",
-        ".tex",
-        ".text",
-        ".tf",
-        ".tfvars",
-        ".thrift",
-        ".toml",
-        ".topojson",
-        ".ts",
-        ".ts.map",
-        ".tsx",
-        ".tsx.map",
-        ".tsv",
-        ".txt",
-        ".v",
-        ".vh",
-        ".vue",
-        ".vtt",
-        ".wxml",
-        ".wxss",
-        ".xml",
-        ".yaml",
-        ".yml",
-        ".zig",
-        ".zsh",
-    }
-)
-TEXT_PREVIEW_FILENAMES = frozenset(
-    {
-        ".babelrc",
-        ".editorconfig",
-        ".env",
-        ".eslintrc",
-        ".gitattributes",
-        ".gitignore",
-        ".gitmodules",
-        ".npmrc",
-        ".prettierrc",
-        ".stylelintrc",
-        ".yarnrc",
-        "containerfile",
-        "dockerfile",
-    }
-)
-PREVIEW_PERMISSION_DENIED_MESSAGE = "Preview unavailable: permission denied"
-PREVIEW_UNSUPPORTED_MESSAGE = "Preview unavailable for this file type"
-PREVIEW_ERROR_MESSAGE = "Preview unavailable"
-IMAGE_PREVIEW_DEPENDENCY_MESSAGE = "Preview unavailable: install `chafa` for image preview"
-GREP_PREVIEW_ERROR_MESSAGE = "Preview unavailable: failed to load context"
-
-# Type alias for grep context cache key
-GrepContextCacheKey = tuple[str, int, int, int, int, int]
+PREVIEW_UNSUPPORTED_MESSAGE = _preview_unsupported_message
+preview_max_bytes_from_kib = _preview_max_bytes_from_kib
 
 
 class BrowserSnapshotLoader(Protocol):
@@ -1055,10 +903,6 @@ def _contains_path(entries, path: str) -> bool:
     return any(entry.path == path for entry in entries)
 
 
-def _normalize_preview_newlines(text: str) -> str:
-    return text.replace("\r\n", "\n")
-
-
 def _normalize_directory_cache_path(path: str) -> str:
     if is_windows_drives_root(path):
         return path
@@ -1067,504 +911,5 @@ def _normalize_directory_cache_path(path: str) -> str:
     if is_windows_path(path):
         return normalize_windows_path(path)
     return str(Path(path).expanduser().resolve())
-
-
-def _build_text_preview_cache_key(
-    path: Path,
-    preview_max_bytes: int,
-    enable_text_preview: bool,
-    enable_image_preview: bool,
-    enable_pdf_preview: bool,
-    enable_office_preview: bool,
-    preview_columns: int,
-) -> tuple[str, int, int, int, bool, bool, bool, bool, int] | "FilePreviewState":
-    preview_limit = max(1, preview_max_bytes)
-    try:
-        stat = path.stat()
-    except PermissionError:
-        return FilePreviewState.permission_denied()
-    except OSError:
-        return FilePreviewState.error()
-    return (
-        str(path),
-        stat.st_mtime_ns,
-        stat.st_size,
-        preview_limit,
-        enable_text_preview,
-        enable_image_preview,
-        enable_pdf_preview,
-        enable_office_preview,
-        max(1, preview_columns),
-    )
-
-
-def _build_grep_context_cache_key(
-    path: Path,
-    line_number: int,
-    context_lines: int,
-    preview_max_bytes: int,
-) -> GrepContextCacheKey | "ContextPreviewState":
-    preview_limit = max(1, preview_max_bytes)
-    try:
-        stat = path.stat()
-    except PermissionError:
-        return ContextPreviewState.with_message(PREVIEW_PERMISSION_DENIED_MESSAGE)
-    except OSError:
-        return ContextPreviewState.with_message(GREP_PREVIEW_ERROR_MESSAGE)
-    return (
-        str(path),
-        stat.st_mtime_ns,
-        stat.st_size,
-        line_number,
-        context_lines,
-        preview_limit,
-    )
-
-
-def _load_text_preview(
-    path: Path,
-    *,
-    preview_max_bytes: int = TEXT_PREVIEW_MAX_BYTES,
-    enable_text_preview: bool = True,
-    enable_image_preview: bool = True,
-    enable_pdf_preview: bool = True,
-    enable_office_preview: bool = True,
-    document_preview_loader: "DocumentPreviewLoader | None" = None,
-    image_preview_loader: "ImagePreviewLoader | None" = None,
-    preview_columns: int = DEFAULT_IMAGE_PREVIEW_COLUMNS,
-) -> "FilePreviewState":
-    if _is_image_preview_candidate(path):
-        if not enable_image_preview:
-            return FilePreviewState.unavailable()
-        loader = image_preview_loader or ChafaImagePreviewLoader()
-        preview = loader.load_preview(path, preview_columns=max(1, preview_columns))
-        if preview is not None:
-            return preview
-        return FilePreviewState.with_message(IMAGE_PREVIEW_DEPENDENCY_MESSAGE)
-
-    if _is_pdf_preview_candidate(path):
-        if not enable_pdf_preview:
-            return FilePreviewState.unavailable()
-        preview = _load_pdf_preview(path, preview_max_bytes=preview_max_bytes)
-        if preview is not None:
-            return preview
-        return FilePreviewState.unsupported()
-
-    if _is_office_preview_candidate(path):
-        if not enable_office_preview:
-            return FilePreviewState.unavailable()
-        loader = document_preview_loader or PandocDocumentPreviewLoader()
-        preview = loader.load_preview(path, preview_max_bytes=preview_max_bytes)
-        if preview is not None:
-            return preview
-        return FilePreviewState.unsupported()
-
-    if not enable_text_preview:
-        return FilePreviewState.unavailable()
-
-    preview_limit = max(1, preview_max_bytes)
-    try:
-        with path.open("rb") as handle:
-            chunk = handle.read(preview_limit + 1)
-    except PermissionError:
-        return FilePreviewState.permission_denied()
-    except OSError:
-        return FilePreviewState.error()
-
-    if b"\x00" in chunk[:preview_limit]:
-        return FilePreviewState.unsupported()
-
-    truncated = len(chunk) > preview_limit
-    preview_bytes = chunk[:preview_limit]
-    try:
-        preview_text = _normalize_preview_newlines(preview_bytes.decode("utf-8"))
-    except UnicodeDecodeError:
-        return FilePreviewState.unsupported()
-
-    return FilePreviewState.with_content(preview_text, truncated)
-
-
-class DocumentPreviewLoader(Protocol):
-    def load_preview(
-        self,
-        path: Path,
-        *,
-        preview_max_bytes: int,
-    ) -> "FilePreviewState | None": ...
-
-
-class ImagePreviewLoader(Protocol):
-    def load_preview(
-        self,
-        path: Path,
-        *,
-        preview_columns: int,
-    ) -> "FilePreviewState | None": ...
-
-
-@dataclass
-class PandocDocumentPreviewLoader:
-    pandoc_path: str | None = field(default=None, init=False, repr=False)
-    pandoc_missing: bool = field(default=False, init=False, repr=False)
-
-    def load_preview(
-        self,
-        path: Path,
-        *,
-        preview_max_bytes: int,
-    ) -> "FilePreviewState | None":
-        pandoc = self._resolve_pandoc()
-        if pandoc is None:
-            return None
-        try:
-            result = subprocess.run(
-                [
-                    pandoc,
-                    "--from",
-                    path.suffix.lstrip(".").lower(),
-                    "--to",
-                    "markdown",
-                    str(path),
-                ],
-                check=True,
-                capture_output=True,
-            )
-        except (OSError, subprocess.SubprocessError, ValueError):
-            return None
-
-        try:
-            content = _normalize_preview_newlines(result.stdout.decode("utf-8"))
-        except UnicodeDecodeError:
-            content = _normalize_preview_newlines(
-                result.stdout.decode("utf-8", errors="ignore")
-            )
-        if not content.strip():
-            return None
-        return _truncate_preview_text(content, preview_max_bytes)
-
-    def _resolve_pandoc(self) -> str | None:
-        if self.pandoc_missing:
-            return None
-        if self.pandoc_path is not None:
-            return self.pandoc_path
-        pandoc = shutil.which("pandoc")
-        if pandoc is None:
-            self.pandoc_missing = True
-            return None
-        self.pandoc_path = pandoc
-        return pandoc
-
-
-@dataclass
-class ChafaImagePreviewLoader:
-    chafa_path: str | None = field(default=None, init=False, repr=False)
-    chafa_missing: bool = field(default=False, init=False, repr=False)
-
-    def load_preview(
-        self,
-        path: Path,
-        *,
-        preview_columns: int,
-    ) -> "FilePreviewState | None":
-        chafa = self._resolve_chafa()
-        if chafa is None:
-            return None
-        try:
-            result = subprocess.run(
-                [
-                    chafa,
-                    "--format",
-                    "symbols",
-                    "--colors",
-                    "full",
-                    "--animate",
-                    "off",
-                    "--fit-width",
-                    "--size",
-                    f"{max(1, preview_columns)}x",
-                    str(path),
-                ],
-                check=True,
-                capture_output=True,
-            )
-        except (OSError, subprocess.SubprocessError, ValueError):
-            return None
-
-        try:
-            content = _normalize_preview_newlines(result.stdout.decode("utf-8"))
-        except UnicodeDecodeError:
-            content = _normalize_preview_newlines(
-                result.stdout.decode("utf-8", errors="ignore")
-            )
-        if not content.strip():
-            return None
-        return FilePreviewState.with_content(content, False, content_kind="image")
-
-    def _resolve_chafa(self) -> str | None:
-        if self.chafa_missing:
-            return None
-        if self.chafa_path is not None:
-            return self.chafa_path
-        chafa = shutil.which("chafa")
-        if chafa is None:
-            self.chafa_missing = True
-            return None
-        self.chafa_path = chafa
-        return chafa
-
-
-def _load_pdf_preview(
-    path: Path,
-    *,
-    preview_max_bytes: int,
-) -> "FilePreviewState | None":
-    pdftotext = shutil.which("pdftotext")
-    if pdftotext is None:
-        return None
-    try:
-        # Windowsではパスをダブルクォートで囲む必要がある
-        path_str = str(path)
-        if " " in path_str:
-            path_str = f'"{path_str}"'
-        result = subprocess.run(
-            [pdftotext, "-q", path_str, "-"],
-            check=True,
-            capture_output=True,
-        )
-    except (OSError, subprocess.SubprocessError, FileNotFoundError) as e:
-        import sys
-        sys.stderr.write(f"PDF preview error: {type(e).__name__}: {e}\n")
-        if hasattr(e, 'stderr') and e.stderr:
-            sys.stderr.write(f"pdftotext stderr: {e.stderr}\n")
-        return None
-    try:
-        content = _normalize_preview_newlines(result.stdout.decode("utf-8"))
-    except UnicodeDecodeError:
-        content = _normalize_preview_newlines(result.stdout.decode("utf-8", errors="ignore"))
-    if not content.strip():
-        # 空のコンテンツは、画像ベースのPDFかテキストレイヤーがないPDF
-        return FilePreviewState.with_message(
-            "PDF preview: no text content found"
-        )
-    return _truncate_preview_text(content, preview_max_bytes)
-
-
-def _load_grep_context_preview(
-    path: Path,
-    line_number: int,
-    context_lines: int,
-    *,
-    preview_max_bytes: int = TEXT_PREVIEW_MAX_BYTES,
-) -> "ContextPreviewState":
-    preview_limit = max(1, preview_max_bytes)
-    start_line = max(1, line_number - max(0, context_lines))
-    end_line = line_number + max(0, context_lines)
-    lines: list[str] = []
-    last_line = 0
-    bytes_read = 0
-    binary_detected = False
-
-    try:
-        # Single file open for both binary detection and context extraction
-        with path.open("rb") as handle:
-            current_line = 0
-            while current_line < end_line:
-                line_bytes = handle.readline()
-                if not line_bytes:
-                    break  # EOF
-
-                bytes_read += len(line_bytes)
-                current_line += 1
-
-                # Binary detection (only check first preview_limit bytes)
-                if not binary_detected and bytes_read <= preview_limit:
-                    if b"\x00" in line_bytes:
-                        return ContextPreviewState.with_message(PREVIEW_UNSUPPORTED_MESSAGE)
-                    try:
-                        line_bytes.decode("utf-8")
-                    except UnicodeDecodeError:
-                        return ContextPreviewState.with_message(PREVIEW_UNSUPPORTED_MESSAGE)
-
-                # Collect context lines
-                if current_line >= start_line:
-                    try:
-                        line_text = _normalize_preview_newlines(line_bytes.decode("utf-8"))
-                        lines.append(line_text)
-                        last_line = current_line
-                    except UnicodeDecodeError:
-                        return ContextPreviewState.with_message(GREP_PREVIEW_ERROR_MESSAGE)
-
-    except PermissionError:
-        return ContextPreviewState.with_message(PREVIEW_PERMISSION_DENIED_MESSAGE)
-    except OSError:
-        return ContextPreviewState.with_message(GREP_PREVIEW_ERROR_MESSAGE)
-
-    if not lines or last_line < line_number:
-        return ContextPreviewState.with_message(GREP_PREVIEW_ERROR_MESSAGE)
-
-    return ContextPreviewState.with_content(
-        "".join(lines),
-        start_line=start_line,
-        highlight_line=line_number,
-    )
-
-
-@dataclass(frozen=True)
-class FilePreviewState:
-    kind: Literal["content", "message", "unavailable"]
-    content: str | None = None
-    content_kind: Literal["text", "image"] = "text"
-    message: str | None = None
-    truncated: bool = False
-
-    @classmethod
-    def with_content(
-        cls,
-        content: str,
-        truncated: bool,
-        *,
-        content_kind: Literal["text", "image"] = "text",
-    ) -> "FilePreviewState":
-        return cls(
-            kind="content",
-            content=content,
-            content_kind=content_kind,
-            truncated=truncated,
-        )
-
-    @classmethod
-    def with_message(cls, message: str) -> "FilePreviewState":
-        return cls(kind="message", message=message)
-
-    @classmethod
-    def permission_denied(cls) -> "FilePreviewState":
-        return cls(kind="message", message=PREVIEW_PERMISSION_DENIED_MESSAGE)
-
-    @classmethod
-    def unsupported(cls) -> "FilePreviewState":
-        return cls(kind="message", message=PREVIEW_UNSUPPORTED_MESSAGE)
-
-    @classmethod
-    def unavailable(cls) -> "FilePreviewState":
-        return cls(kind="unavailable")
-
-    @classmethod
-    def error(cls) -> "FilePreviewState":
-        return cls(kind="message", message=PREVIEW_ERROR_MESSAGE)
-
-
-@dataclass(frozen=True)
-class ContextPreviewState:
-    content: str | None = None
-    message: str | None = None
-    start_line: int | None = None
-    highlight_line: int | None = None
-
-    @classmethod
-    def with_content(
-        cls,
-        content: str,
-        *,
-        start_line: int,
-        highlight_line: int,
-    ) -> "ContextPreviewState":
-        return cls(
-            content=content,
-            start_line=start_line,
-            highlight_line=highlight_line,
-        )
-
-    @classmethod
-    def with_message(cls, message: str) -> "ContextPreviewState":
-        return cls(message=message)
-
-
 def _format_grep_preview_title(result: GrepSearchResultState) -> str:
     return f"Preview: {result.display_path}:{result.line_number}"
-
-
-def _is_text_content(path: Path, blocksize: int = 512) -> bool:
-    """ファイル内容からテキストかどうかをヒューリスティックに判定.
-
-    Args:
-        path: 判定対象のファイルパス
-        blocksize: 読み込むバイト数（デフォルト512）
-
-    Returns:
-        テキストと判定された場合は True、バイナリと判定された場合は False
-    """
-    try:
-        with path.open("rb") as f:
-            chunk = f.read(blocksize)
-    except (PermissionError, OSError):
-        return False
-
-    # 空ファイルはテキスト扱い
-    if not chunk:
-        return True
-
-    # 1. NULLバイトチェック（即バイナリ判定）
-    if b"\x00" in chunk:
-        return False
-
-    # 2. UTF-8として読めるならOK（高速）
-    try:
-        chunk.decode("utf-8")
-        return True
-    except UnicodeDecodeError:
-        pass
-
-    # 3. printable率で雑に判定（軽量）
-    # ASCII印刷可能文字（32-126）、タブ（9）、LF（10）、CR（13）をカウント
-    printable = sum(
-        (32 <= b <= 126) or b in (9, 10, 13)
-        for b in chunk
-    )
-    return printable / len(chunk) > 0.7
-
-
-def _is_preview_candidate(path: Path) -> bool:
-    if (
-        _is_image_preview_candidate(path)
-        or _is_pdf_preview_candidate(path)
-        or _is_office_preview_candidate(path)
-    ):
-        return True
-
-    # 既存：拡張子ベースの判定（高速パス）
-    if path.name.casefold() in TEXT_PREVIEW_FILENAMES:
-        return True
-    suffix = path.suffix.casefold()
-    if suffix in TEXT_PREVIEW_EXTENSIONS:
-        return True
-
-    # 新規：拡張子がない、またはリストにないファイルはヒューリスティック判定
-    return _is_text_content(path)
-
-
-def _is_pdf_preview_candidate(path: Path) -> bool:
-    return path.suffix.casefold() in PDF_PREVIEW_EXTENSIONS
-
-
-def _is_image_preview_candidate(path: Path) -> bool:
-    return path.suffix.casefold() in IMAGE_PREVIEW_EXTENSIONS
-
-
-def _is_office_preview_candidate(path: Path) -> bool:
-    return path.suffix.casefold() in OFFICE_PREVIEW_EXTENSIONS
-
-
-def _truncate_preview_text(content: str, preview_max_bytes: int) -> FilePreviewState:
-    preview_limit = max(1, preview_max_bytes)
-    encoded = content.encode("utf-8")
-    truncated = len(encoded) > preview_limit
-    if not truncated:
-        return FilePreviewState.with_content(content, False)
-
-    preview_bytes = encoded[:preview_limit]
-    preview_text = preview_bytes.decode("utf-8", errors="ignore")
-    return FilePreviewState.with_content(preview_text, True)
-
-
-def preview_max_bytes_from_kib(preview_max_kib: int) -> int:
-    return max(1, preview_max_kib) * 1024
